loepenr filter

+++ src/main/kotlin/no/nav/helsearbeidsgiver/soeknad/SykepengesoeknadFilter.kt
@@ -14,7 +14,7 @@ data class SykepengesoeknadFilter( // TODO: Kanskje denne kan de-dupliseres / ge
     val fnr: String? = null,
     val fom: LocalDate? = null,
     val tom: LocalDate? = null,
-    val fraLoepenr: Int? = null,
+    val fraLoepenr: ULong? = null,
 ) {
     init {
         require(erGyldig(orgnr))
